gfx/layers/apz/test/mochitest/test_group_keyboard.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Various keyboard scrolling tests</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="apz_test_utils.js"></script>
  <script type="application/javascript" src="apz_test_native_event_utils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

var subtests = [
  {"file": "helper_key_scroll.html", prefs: [["apz.test.logging_enabled", true],
                                             ["test.events.async.enabled", true]]},
  {"file": "helper_bug1674935.html", prefs: []},
  {"file": "helper_bug1719330.html", prefs: [["general.smoothScroll", false]]},
  {"file": "helper_bug1695598.html"},
  {"file": "helper_scroll_snap_on_page_down_scroll.html"},
  {"file": "helper_scroll_snap_on_page_down_scroll.html",
    prefs: [["test.events.async.enabled", true]]},
  {"file": "helper_transform_end_on_keyboard_scroll.html",
    prefs: [["general.smoothScroll", false]] },
];
subtests.push(...buildRelativeScrollSmoothnessVariants("key", ["scrollBy"]));
subtests.push(...buildRelativeScrollSmoothnessVariants("native-key", ["scrollBy", "scrollTo", "scrollTop"]));

if (isKeyApzEnabled()) {
  SimpleTest.waitForExplicitFinish();
  window.onload = function() {
    runSubtestsSeriallyInFreshWindows(subtests)
    .then(SimpleTest.finish, SimpleTest.finishWithFailure);
  };
} else {
  SimpleTest.ok(true, "Keyboard APZ is disabled");
}
  </script>
</head>
<body>
  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1383365">Async key scrolling test</a>
</body>
</html>
